vue3: always extract argTypes, even without parameters.docs
