+++ code/renderers/vue3/src/entry-preview.ts
@@ -1,4 +1,33 @@
-export const parameters: {} = { renderer: 'vue3' };
+import {
+  type ArgTypesExtractor,
+  enhanceArgTypes,
+  extractComponentDescription,
+} from 'storybook/internal/docs-tools';
+import type { ArgTypesEnhancer } from 'storybook/internal/types';
+
+import { extractArgTypes } from './extractArgTypes';
+import type { VueRenderer } from './types';
+
 export { render, renderToCanvas } from './render';
 export { decorateStory as applyDecorators } from './decorateStory';
 export { mount } from './mount';
+
+export const parameters: {
+  renderer: 'vue3';
+  docs: {
+    story: {
+      inline: boolean;
+    };
+    extractArgTypes: ArgTypesExtractor;
+    extractComponentDescription: (component?: any) => string;
+  };
+} = {
+  renderer: 'vue3',
+  docs: {
+    story: { inline: true },
+    extractArgTypes,
+    extractComponentDescription,
+  },
+};
+
+export const argTypesEnhancers: ArgTypesEnhancer<VueRenderer>[] = [enhanceArgTypes];
